requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==3.2.16
exceptiongroup==1.1.0
importlib-metadata==6.0.0
iniconfig==2.0.0
linecache2==1.0.0
packaging==22.0
Pillow==9.4.0
pluggy==1.0.0
pytest==7.2.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
traceback2==1.4.0
typing_extensions==4.4.0
unittest2==1.1.0
zipp==3.11.0
